LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/document-write/svg-script-href.tentative.sub-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_not_equals: speculative case did not fetch got disallowed value ""

    <!-- speculative case in document.write -->
    <svg><script href="/html/syntax/speculative-parsing/resources/stash.py?action=put&amp;uuid=7bf7fc81-75f6-4954-ba4d-fe05d1e6e82d&amp;encodingcheck=&Gbreve;"></script></svg>


FAIL Speculative parsing, document.write(): svg-script-href Unhandled rejection: assert_not_equals: speculative case did not fetch got disallowed value ""
